updating integration tests and passing model option to truncate utility function

+++ __tests__/util/truncate.js
@@ -1,12 +1,14 @@
 import database from '../../src/database';
 
-export default async function truncate() {
+export default async function truncate(choosedModel) {
   return await Promise.all(
     Object.keys(database.connection.models).map(key => {
-      return database.connection.models[key].destroy({
-        truncate: true,
-        force: true,
-      });
+      if (key === 'User' || key === choosedModel){
+        return database.connection.models[key].destroy({
+          truncate: true,
+          force: true,
+        });
+      }
     })
   );
 }
